Update nokogiri to 1.8.5
We've got a Github vulnerability alert:

CVE-2018-14404
moderate severity

Vulnerable versions: < 1.8.5
Patched version: 1.8.5

A NULL pointer dereference vulnerability exists in the
xpath.c:xmlXPathCompOpEval() function of libxml2 through 2.9.8 when parsing an
invalid XPath expression in the XPATH_OP_AND or XPATH_OP_OR case. Applications
processing untrusted XSL format inputs with the use of the libxml2 library may
be vulnerable to a denial of service attack due to a crash of the application.